Assignment 11: Authorization -- Complete

+++ src/db/queries.posts.js
@@ -1,5 +1,6 @@
 const Post = require("./models").Post;
 const Topic = require("./models").Topic;
+const Authorizer = require("../policies/post");
 
 module.exports = {
   addPost(newPost, callback) {
@@ -20,32 +21,48 @@ module.exports = {
       callback(err);
     });
   },
-  deletePost(id, callback) {
-    return Post.destroy({
-      where: { id }
+  deletePost(req, callback) {
+    return Post.findById(req.params.id)
+    .then( post => {
+      const authorized = new Authorizer(req.user, post).destroy();
+      if (authorized) {
+        Post.destroy({
+          where: { id: req.params.id }
+        })
+        .then( deletedRecordsCount => {
+          callback(null, deletedRecordsCount);
+        })
+        .catch((err) => {
+          callback(err);
+        });
+      } else {
+        req.flash("notice", "You are not authorized to do that.")
+        callback(401);
+      }
     })
-    .then( deletedRecordsCount => {
-      callback(null, deletedRecordsCount);
-    })
-    .catch((err) => {
-      callback(err);
-    });
   },
-  updatePost(id, updatedPost, callback) {
-    return Post.findById(id)
+  updatePost(req, updatedPost, callback) {
+    
+    return Post.findById(req.params.id)
     .then( post => {
       if (!post) {
         return callback("Post not found");
       }
-      post.update(updatedPost, {
-        fields: Object.keys(updatedPost)
-      })
-      .then( () => {
-        callback(null, post);
-      })
-      .catch( err => {
-        callback(err);
-      });
+      const authorized = new Authorizer(req.user, post).update();
+      if (authorized) {
+        post.update(updatedPost, {
+          fields: Object.keys(updatedPost)
+        })
+        .then( () => {
+          callback(null, post);
+        })
+        .catch( err => {
+          callback(err);
+        });
+      } else {
+        req.flash("notice", "You are not authorized to do that.")
+        callback(401);
+      }
     });
   }
 }
